+++ packages/epetra/src/Epetra_HashTable.h
@@ -53,11 +53,11 @@ class Epetra_HashTable : public Epetra_Object
 {
   struct Node
   {
-     int Key;
+     long long Key;
      int Value;
      Node * Ptr;
 
-     Node( const int key = 0, const int value = 0, Node * ptr = 0 )
+     Node( const long long key = 0, const int value = 0, Node * ptr = 0 )
      : Key(key), Value(value), Ptr(ptr) {}
 
     private:
@@ -72,7 +72,7 @@ class Epetra_HashTable : public Epetra_Object
   int Size_;
   unsigned int Seed_;
 
-  int Func( const int key ) { return (Seed_ ^ key)%Size_; }
+  int Func( const long long key ) { return (int) ((Seed_ ^ key)%Size_); }
      
  public:
 
@@ -115,14 +115,14 @@ class Epetra_HashTable : public Epetra_Object
     delete [] Container_;
   }
 
-  void Add( const int key, const int value )
+  void Add( const long long key, const int value )
   {
     int v = Func(key);
     Node * n1 = Container_[v];
     Container_[v] = new Node(key,value,n1);
   }
 
-  int Get( const int key )
+  int Get( const long long key )
   {
     Node * n = Container_[ Func(key) ];
     while( n && (n->Key != key) ) n = n->Ptr;
